Task Manager v. 0.2.2 -Перенести класс Node в InMemoryHistoryManager -Сделать метод removeNode класса InMemoryHistoryManager приватным
